requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
flask-restplus==0.13.0
flask-restx==1.3.0
importlib_metadata==8.2.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.20.0
